+++ yaqc_bluesky/_device.py
@@ -6,10 +6,12 @@ import yaqc  # type: ignore
 from ._has_position import HasPosition
 from ._is_sensor import IsSensor
 from ._has_measure_trigger import HasMeasureTrigger
+from ._has_mapping import HasMapping
 
 
 traits = [
     ("has-position", HasPosition),
+    ("has-mapping", HasMapping),
     ("has-measure-trigger", HasMeasureTrigger),
     ("is-sensor", IsSensor),
 ]