+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+## [v2021.9.0]
+
 ### Changed
 - state of busy no longer returned when devices are read
 - happi name field no longer overloaded, upstream now allows short names
@@ -45,7 +47,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ### Added
 - initial release
 
-[Unreleased]: https://github.com/bluesky/yaqc-bluesky/compare/v2021.1.1...HEAD
+[Unreleased]: https://github.com/bluesky/yaqc-bluesky/compare/v2021.9.0...HEAD
+[2021.9.0]: https://github.com/bluesky/yaqc-bluesky/compare/v2020.1.1...v2021.9.0
 [2021.1.1]: https://github.com/bluesky/yaqc-bluesky/compare/v2020.1.0...v2021.1.1
 [2021.1.0]: https://github.com/bluesky/yaqc-bluesky/compare/v2020.07.1...v2021.1.0
 [2020.07.1]: https://github.com/bluesky/yaqc-bluesky/compare/v2020.07.0...v2020.07.1